requirements.txt
compressai==1.2.8
nvidia-ml-py @ file:///home/conda/feedstock_root/build_artifacts/nvidia-ml-py_1769147335600/work
PyTurboJPEG @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyturbojpeg_1771658241/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1770223234623/work
setproctitle==1.3.7
timm==1.0.24
torch==2.6.0+cu126
torchinfo==1.8.0
torchvision==0.21.0+cu126
jsonargparse==4.40.2
tensorboard==2.20.0